4.20+4.21

+++ controllers/blogs.js
@@ -3,14 +3,14 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
-const getTokenFrom = request => {
+/*const getTokenFrom = request => {
   const authorization = request.get('authorization')
   if (authorization && authorization.toLowerCase().startsWith('bearer')) {
     return authorization.substring(7)
   }
   return null
 }
-
+*/
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({})
   response.json(blogs.map(blog => blog.toJSON()))
@@ -29,13 +29,12 @@ blogsRouter.get('/:id', (request, response) => {
 })
 
 blogsRouter.post('/', async (request, response, next) => {
-
   const body = request.body
-  const token = getTokenFrom(request)
+  /*const token = getTokenFrom(request)*/
 
   try {
-    const decodedToken = jwt.verify(token, process.env.SECRET)
-    if (!token || !decodedToken.id) {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
       return response.status(401).json({ error: 'Token missing or invalid'})
     }
     const user = await User.findById(decodedToken.id)
@@ -64,9 +63,23 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const blog = await Blog.findById(request.params.id)
+
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'Token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    if (blog.user.id.toString() === user._id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } else {
+      response.status(401).json({ error: 'Token missing or invalid' })
+    }
   } catch (exception) {
     next(exception)
   }
